Se hiz algunas modificaciones para los endpoints de crear usuario y se creo otros endpoints para cambiar contraseña se manda un codigo al correo del user que desea cambiar su contraseña

+++ src/main/java/com/plataforma/repository/PersonR.java
@@ -7,7 +7,7 @@ import java.util.List;
 public interface PersonR {
     List<Person> findAll();
     Person getById(Long id);
-    Integer save(Person obj);
+    Long save(Person obj);
     boolean updateById(Person obj, Long id);
     boolean deleteById(Long id);
 }
